requirements.txt
pyodbc
mysql-connector 
pyyaml
